+++ src/features/help.ts
@@ -1,4 +1,4 @@
-import { t } from "i18next";
+import { changeLanguage, t } from "i18next";
 import { TextMessageCtx } from "types/context";
 import { upsertPrismaChat } from "utils/prisma";
 import { isCleanCommand } from "utils/telegraf";
@@ -7,28 +7,20 @@ export class Help {
   /**
    * Shows help message
    * @param ctx Text message context
+   * @param cleanCommand Clean command name
    */
-  public async command(ctx: TextMessageCtx): Promise<void> {
-    if (!isCleanCommand("help", ctx.message.text) && !isCleanCommand("start", ctx.message.text)) {
+  public async command(ctx: TextMessageCtx, cleanCommand: string): Promise<void> {
+    if (!isCleanCommand(cleanCommand, ctx.message.text)) {
       return; // Not clean command, ignore.
     }
 
-    const { language: lng } = await upsertPrismaChat(ctx.chat, ctx.message.from);
+    const { language } = await upsertPrismaChat(ctx.chat, ctx.message.from);
+    await changeLanguage(language);
 
-    const featureList = [
-      t("addingBots:help", { lng }),
-      t("language:help", { lng }),
-      t("profanityFilter:help", { lng }),
-      t("voteban:help", { lng }),
-    ]
-      .sort((a, b) => a.localeCompare(b))
-      .join("\n\n");
-    const msg = t("common:help", { BOT_LINK: `tg:user?id=${ctx.botInfo.id}`, FEATURE_LIST: featureList, lng });
+    const msg = t("common:help", { BOT_LINK: `tg:user?id=${ctx.botInfo.id}` });
+    const replyToMessageId = ctx.chat.type === "private" ? undefined : ctx.message.message_id;
 
-    await ctx.reply(msg, {
-      parse_mode: "HTML",
-      reply_to_message_id: ctx.chat.type === "private" ? undefined : ctx.message.message_id,
-    });
+    await ctx.reply(msg, { parse_mode: "HTML", reply_to_message_id: replyToMessageId });
   }
 }