+++ src/features/timeZone.ts
@@ -1,6 +1,6 @@
 import { format, getTimezoneOffset } from "date-fns-tz";
 import { settings, SettingsAction } from "features/settings";
-import { t } from "i18next";
+import { changeLanguage, t } from "i18next";
 import { InlineKeyboardButton } from "telegraf/typings/core/types/typegram";
 import { CallbackCtx } from "types/context";
 import { PAGE_SIZE } from "utils/consts";
@@ -14,36 +14,40 @@ export class TimeZone {
    * @param chatId Id of the chat which is edited
    * @param skip Skip count
    */
-  public async renderSettings(ctx: CallbackCtx, chatId: number, skip: number): Promise<void> {
+  public async renderSettings(ctx: CallbackCtx, chatId: number, skip?: number): Promise<void> {
     if (!ctx.chat || isNaN(chatId)) {
-      return; // Something went wrong
+      throw new Error("Chat is not defined to render time zone settings.");
     }
 
-    const { language: lng } = await upsertPrismaChat(ctx.chat, ctx.callbackQuery.from);
-    const prismaChat = await settings.resolvePrismaChat(ctx, chatId, lng);
+    const { language } = await upsertPrismaChat(ctx.chat, ctx.callbackQuery.from);
+    await changeLanguage(language);
+    const prismaChat = await settings.resolvePrismaChat(ctx, chatId);
     if (!prismaChat) {
       return; // The user is no longer an administrator, or the bot has been banned from the chat.
     }
 
     const timeZones = this.getAllTimeZones();
     const count = timeZones.length;
+    const valueIndex = timeZones.indexOf(prismaChat.timeZone);
+    // Use provided skip or the index of the current value. Use 0 as the last fallback.
+    const patchedSkip = skip ?? (valueIndex > -1 ? valueIndex : 0);
     const value = `${format(new Date(), "O", { timeZone: prismaChat.timeZone })} ${prismaChat.timeZone}`;
-    const msg = t("timeZone:select", { CHAT_TITLE: prismaChat.displayTitle, VALUE: value, lng });
+    const msg = t("timeZone:select", { CHAT_TITLE: prismaChat.displayTitle, VALUE: value });
 
     await Promise.all([
       ctx.answerCbQuery(),
-      ctx.editMessageText(joinModifiedInfo(msg, { lng, prismaChat, settingName: "timeZone" }), {
+      ctx.editMessageText(joinModifiedInfo(msg, "timeZone", prismaChat), {
         parse_mode: "HTML",
         reply_markup: {
           inline_keyboard: [
-            ...timeZones.slice(skip, skip + PAGE_SIZE).map((tz): InlineKeyboardButton[] => [
+            ...timeZones.slice(patchedSkip, patchedSkip + PAGE_SIZE).map((tz): InlineKeyboardButton[] => [
               {
                 callback_data: `${SettingsAction.TimeZoneSave}?chatId=${chatId}&v=${tz}`,
                 text: `${format(new Date(), "O", { timeZone: tz })} ${tz}`,
               },
             ]),
-            getPagination(`${SettingsAction.TimeZone}?chatId=${chatId}`, { count, skip, take: PAGE_SIZE }),
-            settings.getBackToFeaturesButton(chatId, lng),
+            getPagination(`${SettingsAction.TimeZone}?chatId=${chatId}`, { count, skip: patchedSkip, take: PAGE_SIZE }),
+            settings.getBackToFeaturesButton(chatId),
           ],
         },
       }),
@@ -58,11 +62,12 @@ export class TimeZone {
    */
   public async saveSettings(ctx: CallbackCtx, chatId: number, value: string | null): Promise<void> {
     if (!ctx.chat || isNaN(chatId)) {
-      return; // Something went wrong
+      throw new Error("Chat is not defined to save time zone settings.");
     }
 
-    const { language: lng } = await upsertPrismaChat(ctx.chat, ctx.callbackQuery.from);
-    const prismaChat = await settings.resolvePrismaChat(ctx, chatId, lng);
+    const { language } = await upsertPrismaChat(ctx.chat, ctx.callbackQuery.from);
+    await changeLanguage(language);
+    const prismaChat = await settings.resolvePrismaChat(ctx, chatId);
     if (!prismaChat) {
       return; // The user is no longer an administrator, or the bot has been banned from the chat.
     }
@@ -74,7 +79,7 @@ export class TimeZone {
       prisma.chat.update({ data: { timeZone }, select: { id: true }, where: { id: chatId } }),
       upsertPrismaChatSettingsHistory(chatId, ctx.callbackQuery.from.id, "timeZone"),
     ]);
-    await Promise.all([settings.notifyChangesSaved(ctx, lng), this.renderSettings(ctx, chatId, skip)]);
+    await Promise.all([settings.notifyChangesSaved(ctx), this.renderSettings(ctx, chatId, skip)]);
   }
 
   /**